all ints to floats

+++ src/LoadLevel.hs
@@ -2,7 +2,7 @@ module Loadlevel where
 
 data Direction = GoLeft | GoRight | GoUp | GoDown deriving (Show, Eq)
 
-type Point = (Int, Int)
+type Point = (Float, Float)
 
 type Tile = (Point, Tiletype)
 
@@ -13,15 +13,15 @@ type Maze = [[Tile]]
 makeMaze :: [String] -> Maze
 makeMaze file = map makeLine (zip [0..] file)
 
-makeLine :: (Int, String) -> [Tile]
+makeLine :: (Float, String) -> [Tile]
 makeLine (y, s) = map (mergeInfo y) (map makeTile (zip [0..] s))
 
-makeTile :: (Int, Char) -> (Int, Tiletype)
+makeTile :: (Float, Char) -> (Float, Tiletype)
 makeTile (x, c) | c == '#' = (x, Wall)
                 | c == '-' = (x, Empty)
                 | c == '+' = (x, Pellet)
                 | c == '%' = (x, SuperPellet)
                 | c == '_' = (x, Barrier)
 
-mergeInfo :: Int -> (Int, Tiletype) -> Tile
+mergeInfo :: Float -> (Float, Tiletype) -> Tile
 mergeInfo y (x, t) = ((x, y), t)
